Rename card-search components/utils to search
The search components and utils now search entries (cards and files),
not just cards, so the `card-search` name is misleading. Rename both
directories — components/card-search -> components/search and
utils/card-search -> utils/search — and update all import paths across
host. Also rename the three tests that mirror the subject and their
QUnit module labels for consistency, and update the matching
memory-baseline.json keys. Mechanical rename only; no behavior changes.
