+++ packages/host/app/components/search/hydratable-card.gts
