+++ packages/host/app/components/search/live-recents-provider.gts
