+++ packages/host/app/components/search/panel.gts
