+++ packages/host/app/components/search/search-bar.gts
