+++ packages/host/app/components/search/search-result-error.gts
