+++ packages/host/app/components/search/section-header.gts
@@ -4,7 +4,7 @@ import Component from '@glimmer/component';
 import { BoxelInput, RealmIcon } from '@cardstack/boxel-ui/components';
 import { eq } from '@cardstack/boxel-ui/helpers';
 
-import type { RealmSectionInfo } from '@cardstack/host/utils/card-search/sections';
+import type { RealmSectionInfo } from '@cardstack/host/utils/search/sections';
 
 import type { ComponentLike } from '@glint/template';
