+++ packages/host/app/templates/render/html.gts
@@ -17,7 +17,7 @@ import {
   type Query,
 } from '@cardstack/runtime-common';
 
-import SearchResults from '@cardstack/host/components/card-search/search-results';
+import SearchResults from '@cardstack/host/components/search/search-results';
 import { getCardCollection } from '@cardstack/host/resources/card-collection';
 import { getCard } from '@cardstack/host/resources/card-resource';
 import type RenderStoreService from '@cardstack/host/services/render-store';
