+++ packages/host/app/utils/search/sections.ts
