+++ packages/host/tests/integration/components/search-results-test.gts
@@ -30,7 +30,7 @@ import {
   type SearchEntryWireQuery,
 } from '@cardstack/runtime-common';
 
-import SearchResults from '@cardstack/host/components/card-search/search-results';
+import SearchResults from '@cardstack/host/components/search/search-results';
 import { getCard } from '@cardstack/host/resources/card-resource';
 import type StoreService from '@cardstack/host/services/store';
